open_spiel/python/examples/opponent_shaping/requirements.txt
wandb
jax
jax[cuda] -f https://storage.googleapis.com/jax-releases/jax_cuda_releases.html
distrax
optax
dm-haiku
rlax
open_spiel